Move map classes into subfolder of engine.
