+++ src/engine/behaviors/player-behavior.ts
@@ -2,7 +2,7 @@ import { MoveByAction } from 'engine/actions/move-by'
 import { Player } from 'engine/player'
 import { Behavior } from 'engine/types'
 
-import { creatureAdjustedCost } from '../ai/path-cost-functions'
+import { creatureAdjustedCost } from '../map/path-cost-functions'
 
 /** Special type of behavior where actions are directly controlled by the player. */
 export const PlayerBehavior = (player: Player): Behavior => (_, world) => {
