+++ src/engine/map/map-symbol.ts
