+++ src/engine/map/map-symbolizer.ts
