+++ src/engine/map/map-utils.ts
